@@ -1,21 +1,19 @@
+import process from 'node:process';
 import test from 'ava';
-import proxyquire from 'proxyquire';
-import clearModule from 'clear-module';
+import esmock from 'esmock';
 
-test.beforeEach(() => {
-	clearModule('.');
-});
+const requireFresh = modulePath => import(`${modulePath}?cacheBust=${Date.now()}`);
 
-test('inside WSL 1', t => {
+test('inside WSL 1', async t => {
 	process.env.__IS_WSL_TEST__ = true;
 
 	const originalPlatform = process.platform;
 	Object.defineProperty(process, 'platform', {value: 'linux'});
 
-	const isWsl = proxyquire('.', {
+	const isWsl = await esmock('./index.js', {
 		fs: {
-			readFileSync: () => 'Linux version 3.4.0-Microsoft (Microsoft@Microsoft.com) (gcc version 4.7 (GCC) ) #1 SMP PREEMPT Wed Dec 31 14:42:53 PST 2014'
-		}
+			readFileSync: () => 'Linux version 3.4.0-Microsoft (Microsoft@Microsoft.com) (gcc version 4.7 (GCC) ) #1 SMP PREEMPT Wed Dec 31 14:42:53 PST 2014',
+		},
 	});
 
 	t.true(isWsl());
@@ -24,16 +22,16 @@ test('inside WSL 1', t => {
 	Object.defineProperty(process, 'platform', {value: originalPlatform});
 });
 
-test('inside WSL 2', t => {
+test('inside WSL 2', async t => {
 	process.env.__IS_WSL_TEST__ = true;
 
 	const originalPlatform = process.platform;
 	Object.defineProperty(process, 'platform', {value: 'linux'});
 
-	const isWsl = proxyquire('.', {
+	const isWsl = await esmock('./index.js', {
 		fs: {
-			readFileSync: () => 'Linux version 4.19.43-microsoft-standard (oe-user@oe-host) (gcc version 7.3.0 (GCC)) #1 SMP Mon May 20 19:35:22 UTC 2019'
-		}
+			readFileSync: () => 'Linux version 4.19.43-microsoft-standard (oe-user@oe-host) (gcc version 7.3.0 (GCC)) #1 SMP Mon May 20 19:35:22 UTC 2019',
+		},
 	});
 
 	t.true(isWsl());
@@ -42,32 +40,33 @@ test('inside WSL 2', t => {
 	Object.defineProperty(process, 'platform', {value: originalPlatform});
 });
 
-test('not inside WSL', t => {
+test('not inside WSL', async t => {
 	process.env.__IS_WSL_TEST__ = true;
 
 	const originalPlatform = process.platform;
 	Object.defineProperty(process, 'platform', {value: 'darwin'});
 
-	const isWsl = require('.');
+	const {default: isWsl} = await requireFresh('./index.js');
+
 	t.false(isWsl());
 
 	delete process.env.__IS_WSL_TEST__;
 	Object.defineProperty(process, 'platform', {value: originalPlatform});
 });
 
-test('not inside WSL, but inside Linux', t => {
+test('not inside WSL, but inside Linux', async t => {
 	process.env.__IS_WSL_TEST__ = true;
 
 	const originalPlatform = process.platform;
 	Object.defineProperty(process, 'platform', {value: 'linux'});
 
-	const isWsl = proxyquire('.', {
+	const isWsl = await esmock('./index.js', {
 		fs: {
-			readFileSync: () => 'Linux version 4.19.43-standard (oe-user@oe-host) (gcc version 7.3.0 (GCC)) #1 SMP Mon May 20 19:35:22 UTC 2019'
+			readFileSync: () => 'Linux version 4.19.43-standard (oe-user@oe-host) (gcc version 7.3.0 (GCC)) #1 SMP Mon May 20 19:35:22 UTC 2019',
 		},
 		os: {
-			release: () => ''
-		}
+			release: () => '',
+		},
 	});
 
 	t.false(isWsl());
@@ -76,20 +75,20 @@ test('not inside WSL, but inside Linux', t => {
 	Object.defineProperty(process, 'platform', {value: originalPlatform});
 });
 
-test('inside WSL, but inside docker', t => {
+test('inside WSL, but inside docker', async t => {
 	process.env.__IS_WSL_TEST__ = true;
 
 	const originalPlatform = process.platform;
 	Object.defineProperty(process, 'platform', {value: 'linux'});
 
-	const isWsl = proxyquire('.', {
+	const isWsl = await esmock('./index.js', {
 		fs: {
-			readFileSync: () => 'Linux version 4.19.43-microsoft-standard (oe-user@oe-host) (gcc version 7.3.0 (GCC)) #1 SMP Mon May 20 19:35:22 UTC 2019'
+			readFileSync: () => 'Linux version 4.19.43-microsoft-standard (oe-user@oe-host) (gcc version 7.3.0 (GCC)) #1 SMP Mon May 20 19:35:22 UTC 2019',
 		},
 		'is-docker': () => true,
 		os: {
-			release: () => 'microsoft'
-		}
+			release: () => 'microsoft',
+		},
 	});
 
 	t.false(isWsl());
